backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pandas==2.2.2
anthropic>=0.40.0
python-dotenv==1.0.1
sqlparse==0.5.0
pydantic==2.7.1
chromadb==0.5.0
sentence-transformers==3.0.1
numpy==1.26.4
